Bruk banneret på kandidatlisteoversikten
