+++ src/kandidat/kandidatlisteoversikt/Kandidatlisteoversikt.module.css
@@ -12,11 +12,18 @@
     padding: 0 1rem;
 }
 
+.venstreKolonne {
+    grid-area: filter;
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+}
+
 .filter {
     grid-area: filter;
 }
 
-.antallKandidatlister {
+.overTabell {
     grid-area: over-tabell;
 }
 
@@ -33,13 +40,13 @@
 @media (min-width: 1280px) {
     .wrapper {
         display: grid;
-        grid-template-columns: 14rem auto;
+        grid-template-columns: 19rem auto;
         grid-template-areas:
             'over-filter over-tabell'
             'filter tabell'
             'under-filter under-tabell';
 
-        gap: 2rem;
+        gap: 1.5rem 3rem;
         padding: 0 2rem;
         max-width: calc(var(--innhold-maksbredde) + 4rem);
     }
