+++ src/kandidat/kandidatlisteoversikt/Kandidatlisteoversikt.tsx
@@ -1,25 +1,28 @@
+import { Button, Heading, Pagination } from '@navikt/ds-react';
+import classNames from 'classnames';
 import React from 'react';
 import { connect } from 'react-redux';
-import classNames from 'classnames';
-import { Heading, Pagination } from '@navikt/ds-react';
 
+import { PlusCircleIcon } from '@navikt/aksel-icons';
 import { KandidatlisteSammendrag } from 'felles/domene/kandidatliste/Kandidatliste';
-import { ListeoversiktAction, ListeoversiktActionType } from './reducer/ListeoversiktAction';
+import Banner from 'felles/komponenter/banner/Banner';
+import { ReactComponent as Piktogram } from 'felles/komponenter/piktogrammer/finn-kandidater.svg';
 import { Nettstatus } from 'felles/nettressurs';
-import { Søkekriterier } from './reducer/listeoversiktReducer';
-import AppState from '../state/AppState';
-import EndreModal from './modaler/EndreModal';
-import Filter, { Stillingsfilter } from './filter/Filter';
-import Header from './header/Header';
 import KandidatlisteAction from '../kandidatliste/reducer/KandidatlisteAction';
 import KandidatlisteActionType from '../kandidatliste/reducer/KandidatlisteActionType';
-import Kandidatlistetabell from './tabell/Kandidatlistetabell';
+import AppState from '../state/AppState';
+import css from './Kandidatlisteoversikt.module.css';
+import Filter, { Stillingsfilter } from './filter/Filter';
+import EndreModal from './modaler/EndreModal';
 import MarkerSomMinModal from './modaler/MarkerSomMinModal';
 import OpprettModal from './modaler/OpprettModal';
 import SlettModal from './modaler/SlettModal';
+import { ListeoversiktAction, ListeoversiktActionType } from './reducer/ListeoversiktAction';
+import { Søkekriterier } from './reducer/listeoversiktReducer';
+import Kandidatlistesøk from './søk/Kandidatlistesøk';
+import Kandidatlistetabell from './tabell/Kandidatlistetabell';
 import TabellBody from './tabell/TabellBody';
 import TabellHeader from './tabell/TabellHeader';
-import css from './Kandidatlisteoversikt.module.css';
 
 const SIDESTØRRELSE = 20;
 
@@ -261,22 +264,33 @@ class Kandidatlisteoversikt extends React.Component<Props> {
                         onClose={this.handleLukkModal}
                     />
                 )}
-                <Header
-                    søkeOrd={søkeOrd}
-                    onSøkeOrdChange={this.onSøkeOrdChange}
-                    onSubmitSøkKandidatlister={this.onSubmitSøkKandidatlister}
-                    nullstillSøk={this.onNullstillSøkClick}
-                    opprettListe={this.onOpprettClick}
-                />
+                <Banner tittel="Kandidatlister" ikon={<Piktogram />}>
+                    <Button
+                        variant="secondary"
+                        icon={<PlusCircleIcon />}
+                        onClick={this.onOpprettClick}
+                    >
+                        Opprett ny
+                    </Button>
+                </Banner>
                 <div className={css.wrapper}>
-                    <Filter
-                        className={css.filter}
-                        søkekriterier={søkekriterier}
-                        onVisMineKandidatlister={this.onVisMineKandidatlister}
-                        onVisAlleKandidatlister={this.onVisAlleKandidatlister}
-                        onFilterChange={this.onFilterChange}
-                    />
-                    <div className={css.antallKandidatlister}>
+                    <div className={css.venstreKolonne}>
+                        <Kandidatlistesøk
+                            søkeOrd={søkeOrd}
+                            onSøkeOrdChange={this.onSøkeOrdChange}
+                            onSubmitSøkKandidatlister={this.onSubmitSøkKandidatlister}
+                            nullstillSøk={this.onNullstillSøkClick}
+                        />
+                        <Filter
+                            className={css.filter}
+                            søkekriterier={søkekriterier}
+                            onVisMineKandidatlister={this.onVisMineKandidatlister}
+                            onVisAlleKandidatlister={this.onVisAlleKandidatlister}
+                            onFilterChange={this.onFilterChange}
+                        />
+                    </div>
+
+                    <div className={css.overTabell}>
                         <Heading level="1" size="medium">
                             {tittel}
                         </Heading>
