+++ src/kandidat/kandidatlisteoversikt/søk/Kandidatlistesøk.module.css
@@ -1,0 +1,4 @@
+.søkeskjema {
+    width: 100%;
+    max-width: 30rem;
+}
