+++ src/kandidat/kandidatlisteoversikt/søk/Kandidatlistesøk.tsx
@@ -1,0 +1,30 @@
+import { Search } from '@navikt/ds-react';
+import { FunctionComponent } from 'react';
+import css from './Kandidatlistesøk.module.css';
+
+type Props = {
+    søkeOrd?: string;
+    onSøkeOrdChange: (event: string) => void;
+    onSubmitSøkKandidatlister?: any;
+    nullstillSøk: () => void;
+};
+
+const Kandidatlistesøk: FunctionComponent<Props> = ({
+    søkeOrd,
+    onSøkeOrdChange,
+    onSubmitSøkKandidatlister,
+    nullstillSøk,
+}) => (
+    <form className={css.søkeskjema} onSubmit={onSubmitSøkKandidatlister}>
+        <Search
+            hideLabel={true}
+            label="Søk etter kandidatlister"
+            placeholder="Søk etter kandidatlister"
+            onChange={onSøkeOrdChange}
+            value={søkeOrd}
+            onClear={nullstillSøk}
+        />
+    </form>
+);
+
+export default Kandidatlistesøk;
